requirements.txt
pymongo>=4.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
schedule>=1.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
